test: add snicallback prioritization test
Depending on the content of a cert, it seems that OpenSSL
will prioritize the root cert over the context returned by
SNICallback. This behavior is unexpected. This PR adds a
test that demonstrates the behavior. Ideally there is a fix,
or a explanation for this behavior we can add to the docs.

Fixes: https://github.com/nodejs/node/issues/54235
PR-URL: https://github.com/nodejs/node/pull/54251