test: skip sha128/256 createHash()/hash() on openssl 3.4.
OpenSSL 3.4 has intentionally broken EVP_DigestFinal for SHAKE128 and
SHAKE256 when OSSL_DIGEST_PARAM_XOFLEN is not set because a) the default
length used weakened them from their maximum strength and b) a static
length does not fully make sense for XOFs (which SHAKE* are).

Unfortunately, while crypto.createHash accepts an option argument that can
be something like `{ outputLength: 128 }`, crypto.hash doesn't offer a
similar API. Therefore there is little choice but to skip the test
completely for shake128 and shake256 on openssl >= 3.4.

PR-URL: https://github.com/nodejs/node/pull/56294

Fixes: https://github.com/nodejs/node/issues/56159
Refs: https://github.com/openssl/openssl/commit/b911fef216d1386210ec24e201d54d709528abb4
Refs: https://github.com/openssl/openssl/commit/ad3f28c5fbd5dcbc763a650313fd666b0e339cca